// src/api-serverless/src/generated/models/CreateNewWaveVotingConfig.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { IntRange } from '../models/IntRange';
import { WaveCreditScope } from '../models/WaveCreditScope';
import { WaveCreditType } from '../models/WaveCreditType';
import { WaveScope } from '../models/WaveScope';
import { HttpFile } from '../http/http';

export class CreateNewWaveVotingConfig {
    'scope': WaveScope;
    'credit_type': WaveCreditType;
    'credit_scope': WaveCreditScope;
    /**
    * Only relevant when credit_type=REP. The REP category which is usable as credit. If not set then all categories will be usable as voting credits.
    */
    'credit_category': string | null;
    /**
    * Only relevant when credit_type=REP
    */
    'creditor_id': string | null;
    /**
    * If true then the votes must be signed by voters.
    */
    'signature_required': boolean;
    'period'?: IntRange;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "scope",
            "baseName": "scope",
            "type": "WaveScope",
            "format": ""
        },
        {
            "name": "credit_type",
            "baseName": "credit_type",
            "type": "WaveCreditType",
            "format": ""
        },
        {
            "name": "credit_scope",
            "baseName": "credit_scope",
            "type": "WaveCreditScope",
            "format": ""
        },
        {
            "name": "credit_category",
            "baseName": "credit_category",
            "type": "string",
            "format": ""
        },
        {
            "name": "creditor_id",
            "baseName": "creditor_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "signature_required",
            "baseName": "signature_required",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "period",
            "baseName": "period",
            "type": "IntRange",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateNewWaveVotingConfig.attributeTypeMap;
    }

    public constructor() {
    }
}