// src/api-serverless/src/generated/models/WaveDropsFeed.ts
/**
 * Seize API
 * This is the API interface description for the most commonly used operations in Seize API.  Some modifying endpoints require an authentication token.   We are in the process of documenting all Seize APIs.   If there is an API that you need, please ping us in Discord and we will aim to prioritize its documentation.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DropWithoutWave } from '../models/DropWithoutWave';
import { WaveMin } from '../models/WaveMin';
import { HttpFile } from '../http/http';

export class WaveDropsFeed {
    'wave': WaveMin;
    'drops': Array<DropWithoutWave>;
    'root_drop'?: DropWithoutWave;
    'trace'?: Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "wave",
            "baseName": "wave",
            "type": "WaveMin",
            "format": ""
        },
        {
            "name": "drops",
            "baseName": "drops",
            "type": "Array<DropWithoutWave>",
            "format": ""
        },
        {
            "name": "root_drop",
            "baseName": "root_drop",
            "type": "DropWithoutWave",
            "format": ""
        },
        {
            "name": "trace",
            "baseName": "trace",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return WaveDropsFeed.attributeTypeMap;
    }

    public constructor() {
    }
}